Create app to export/import USDA data.

+++ MSUDTrack/MSUDTrack.DataModels/Models/Record.cs
@@ -27,16 +27,16 @@ namespace MSUDTrack.DataModels.Models
         /// <summary>
         /// The amount of protein this food contains, in grams.
         /// </summary>
-        public int? ProteinGrams { get; set; }
+        public double? ProteinGrams { get; set; }
 
         /// <summary>
         /// The amount of leucine this food contains, in milligrams.
         /// </summary>
-        public int? LeucineMilligrams { get; set; }
+        public double? LeucineMilligrams { get; set; }
 
         /// <summary>
         /// The weight of the food consumed, in grams.
         /// </summary>
-        public int? WeightGrams { get; set; }
+        public double? WeightGrams { get; set; }
     }
 }
